requirements.txt
fastapi
uvicorn[standard]
streamlit
plotly
pandas
duckdb
requests
snowflake-connector-python
pydantic
python-multipart
pyjwt
passlib[bcrypt]
bcrypt<4

openai<1
psycopg2-binary
pymysql
